analytics/requirements.txt
# RabbitMQ
pika==1.3.2

# Variáveis de ambiente
python-dotenv==1.0.1

# ORM e banco de dados PostgreSQL
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Tipagem e validação futura (opcional, mas útil)
pydantic==2.6.4

# Migrations
alembic==1.13.1  

# FastAPI e servidor
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Pydantic Settings para configuração
pydantic-settings==2.2.1

# Para formulários e multipart
python-multipart==0.0.9

# Log melhorado
loguru==0.7.2

# Machine Learning e Data Science
pandas==2.2.2
scikit-learn==1.4.2

# Task queue
celery[rabbitmq]==5.3.6
